take out array of data to the new file
